+++ requirements.dev.txt
@@ -1,10 +1,8 @@
 # development libraries
-tox
-tox-envreport
-tox-pytest-summary
 coverage
 pytest
 pytest-bdd
 pytest-cov
 pytest-subtests
-
+tox
+tox-pytest-summary
