--- setup.cfg
@@ -1,16 +1,0 @@
-[tool:pytest]
-addopts =
-    --cov=ted_sws
-    --cov-report=html
-    --cov-report=xml
-    -r a
-    -q
-    -v
-    --junitxml=junit_report.xml
-
-testpaths = tests
-
-junit_family=xunit2
-
-env_files =
-    .env
