+++ tests/features/metadata_normaliser/__init__.py
@@ -1,8 +1,8 @@
 #!/usr/bin/python3
 
 # __init__.py
-# Date:  07/02/2022
+# Date:  03/02/2022
 # Author: Eugeniu Costetchi
-# Email: costezki.eugen@gmail.com 
+# Email: costezki.eugen@gmail.com
 
 """ """
