--- tests/step_defs/conftest.py
@@ -1,8 +1,0 @@
-#!/usr/bin/python3
-
-# conftest.py
-# Date:  07/02/2022
-# Author: Eugeniu Costetchi
-# Email: costezki.eugen@gmail.com 
-
-""" """
