+++ tests/unit/domain/model/test_notice_creation.py
@@ -49,9 +49,11 @@ def test_notice_invalid_creation():
 
 
 def test_notice_status_validation(publicly_available_notice):
-    with pytest.raises(Exception):
-        publicly_available_notice.update_status_to(NoticeStatus.TRANSFORMED)
-        pprint(publicly_available_notice.dict())
-        pprint(publicly_available_notice.dict().keys())
-        publicly_available_notice._status = NoticeStatus.FAULTY_PACKAGE
-        pprint(publicly_available_notice.dict())
+    publicly_available_notice.update_status_to(NoticeStatus.TRANSFORMED)
+    # pprint(publicly_available_notice.dict())
+    # pprint(publicly_available_notice.dict().keys())
+    publicly_available_notice._status = NoticeStatus.FAULTY_PACKAGE
+    # pprint(publicly_available_notice.dict())
+
+    assert "status" in publicly_available_notice.dict().keys()
+    assert "_status" not in publicly_available_notice.dict().keys()
