+++ tests/unit/notice_fetcher/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python3
+
+# __init__.py
+# Date:  03/02/2022
+# Author: Eugeniu Costetchi
+# Email: costezki.eugen@gmail.com
+
+""" """
