+++ tox.ini
@@ -1,0 +1,73 @@
+[tox]
+# skipping the setup.py installation/requirement for the moment
+skipsdist = true
+envlist =
+#py37,py38,py39,py310 # skipping multiple python distros for the moment
+    py38
+
+[testenv]
+description = Running tests in the corresponding environment
+deps =
+    -rrequirements.txt
+    -rrequirements.dev.txt
+
+commands =
+    {envpython} -m pytest
+
+[testenv:unit]
+description = Running UNIT tests in the corresponding environment
+commands =
+    {envpython} -m pytest tests/unit
+
+[testenv:features]
+commands =
+    {envpython} -m pytest tests/features
+
+[testenv:e2e]
+commands =
+    {envpython} -m pytest tests/e2e
+
+[testenv:cover]
+basepython =
+    python3.8
+commands =
+    {envpython} -m pytest \
+    --cov-report term \
+    --cov-report html \
+    --cov
+
+[pytest]
+log_cli = true
+log_cli_format = %(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
+log_cli_date_format = %Y-%m-%d %H:%M:%S
+
+
+addopts =
+    --cov=ted_sws
+    --cov-report=html
+    --cov-report=term
+    -r a
+    -q
+    -v
+    -s
+    --gherkin-terminal-reporter
+    --disable-pytest-warnings
+    --disable-warnings
+
+
+testpaths =
+    tests
+
+[testenv:linter]
+deps =
+    flake8
+    flake8-bugbear
+    flake8-builtins
+    flake8-comprehensions
+    flake8-string-format
+    black
+    pydocstyle
+commands =
+    flake8
+    black --check .
+    pydocstyle
