--- tox.ini.unused
@@ -1,29 +1,0 @@
-[tox]
-envlist =
-    #py37,py38,py39,py310
-    py38
-
-[testenv]
-deps =
-    -rrequirements.txt
-    -rrequirements.dev.txt
-
-commands =
-    {envpython} setup.py clean --all
-    {envpython} setup.py build
-    {envpython} -m pytest
-
-[testenv:cover]
-basepython =
-    python3.8
-commands =
-    {envpython} -m pytest \
-		--cov-report term \
-		--cov-report html \
-		--cov
-
-[pytest]
-# log_cli = true
-# log_cli_level = DEBUG
-log_cli_format = %(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
-log_cli_date_format = %Y-%m-%d %H:%M:%S
